requirements.txt
PySide6>=6.5.0
openai>=1.0.0
anthropic>=0.7.0
edge-tts>=6.0.0
pyautogui>=0.9.54
keyboard>=0.13.5
pyyaml>=6.0
psycopg2-binary>=2.9.0

